requirements/base.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Celery
celery==5.3.4
redis==5.0.1

# Image processing
Pillow==10.1.0

# Environment variables
python-decouple==3.8

# API documentation
drf-yasg==1.21.7

# Timezone
pytz==2023.3

# Utilities
python-dateutil==2.8.2
